+++ src/main/java/br/com/swconsultoria/nfe/dom/ConfiguracoesNfe.java
@@ -109,7 +109,7 @@ public class ConfiguracoesNfe {
 
         if (Logger.getLogger("").isLoggable(Level.SEVERE)) {
             System.err.println("####################################################################");
-            System.err.println("       Api Java Nfe - Versão 4.00.27 - 26/11/2023");
+            System.err.println("       Api Java Nfe - Versão 4.00.28 - 20/01/2024");
             if (Logger.getLogger("").isLoggable(Level.WARNING)) {
                 System.err.println(" Samuel Olivera - samuel@swconsultoria.com.br ");
             }