+++ src/main/java/br/com/swconsultoria/nfe/schema/eventoSuframaVistoria/X509DataType.java
@@ -1,0 +1,60 @@
+
+package br.com.swconsultoria.nfe.schema.eventoSuframaVistoria;
+
+import javax.xml.bind.annotation.XmlAccessType;
+import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
+import javax.xml.bind.annotation.XmlType;
+
+
+/**
+ * <p>Classe Java de X509DataType complex type.
+ * 
+ * <p>O seguinte fragmento do esquema especifica o conteúdo esperado contido dentro desta classe.
+ * 
+ * <pre>
+ * &lt;complexType name="X509DataType"&gt;
+ *   &lt;complexContent&gt;
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
+ *       &lt;sequence&gt;
+ *         &lt;element name="X509Certificate" type="{http://www.w3.org/2001/XMLSchema}base64Binary"/&gt;
+ *       &lt;/sequence&gt;
+ *     &lt;/restriction&gt;
+ *   &lt;/complexContent&gt;
+ * &lt;/complexType&gt;
+ * </pre>
+ * 
+ * 
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "X509DataType", namespace = "http://www.w3.org/2000/09/xmldsig#", propOrder = {
+    "x509Certificate"
+})
+public class X509DataType {
+
+    @XmlElement(name = "X509Certificate", namespace = "http://www.w3.org/2000/09/xmldsig#", required = true)
+    protected byte[] x509Certificate;
+
+    /**
+     * Obtém o valor da propriedade x509Certificate.
+     * 
+     * @return
+     *     possible object is
+     *     byte[]
+     */
+    public byte[] getX509Certificate() {
+        return x509Certificate;
+    }
+
+    /**
+     * Define o valor da propriedade x509Certificate.
+     * 
+     * @param value
+     *     allowed object is
+     *     byte[]
+     */
+    public void setX509Certificate(byte[] value) {
+        this.x509Certificate = value;
+    }
+
+}